requirements.txt
# FastAPI and dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
supabase==1.0.3
asyncpg==0.27.0
sqlalchemy==2.0.9

# HTTP clients
httpx==0.23.3
aiohttp==3.8.4

# External APIs
openai==0.27.4  # For GPT-4
google-api-python-client==2.79.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
serpapi==0.1.3

# Utilities
tenacity==8.2.2  # For retries
python-dateutil==2.8.2
ujson==5.7.0
validators==0.22.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0

# Development tools
black==23.3.0
isort==5.12.0
mypy==1.2.0
flake8==6.0.0

# Deployment
gunicorn==20.1.0
